move HelpLinks to FileEditor
