+++ catalog/app/containers/Bucket/PackageDialog/DialogError.tsx
@@ -3,10 +3,10 @@ import * as React from 'react'
 import * as M from '@material-ui/core'
 import { fade } from '@material-ui/core/styles'
 
-import StyledLink from 'utils/StyledLink'
+import { WorkflowsConfigLink } from 'components/FileEditor/HelpLinks'
 import { docs } from 'constants/urls'
+import StyledLink from 'utils/StyledLink'
 
-import { WorkflowsConfigLink } from '../HelpLinks'
 import * as ERRORS from '../errors'
 
 const useStyles = M.makeStyles((t) => ({
